requirements.txt
streamlit == 1.2.0
pandas == 1.3.4
numpy == 1.21.4
matplotlib.pyplot
seaborn == 0.11.2